refactor(terraform): rename `common.ts` to `helpers.ts`
Align naming with pulumi plugin.
