+++ plugins/terraform/commands.ts
@@ -10,7 +10,7 @@ import chalk from "chalk"
 import { terraform } from "./cli"
 import { TerraformProvider } from "./provider"
 import { ConfigurationError, ParameterError } from "@garden-io/sdk/exceptions"
-import { prepareVariables, setWorkspace, tfValidate } from "./common"
+import { prepareVariables, setWorkspace, tfValidate } from "./helpers"
 import { ConfigGraph, PluginCommand, PluginCommandParams } from "@garden-io/sdk/types"
 import { join } from "path"
 import { remove } from "fs-extra"
