+++ plugins/terraform/module.ts
@@ -12,7 +12,7 @@ import { joi } from "@garden-io/core/build/src/config/common"
 import { GardenModule, ModuleActionHandlers } from "@garden-io/sdk/types"
 import { ConfigurationError } from "@garden-io/sdk/exceptions"
 import { dependenciesSchema } from "@garden-io/core/build/src/config/service"
-import { TerraformBaseSpec } from "./common"
+import { TerraformBaseSpec } from "./helpers"
 import { TerraformProvider } from "./provider"
 import { baseBuildSpecSchema } from "@garden-io/core/build/src/config/module"
 import { terraformDeploySchemaKeys } from "./action"
