+++ plugins/terraform/test/common.ts
@@ -12,7 +12,7 @@ import { gardenPlugin } from ".."
 import { TerraformProvider } from "../provider"
 import { makeTestGarden, TestGarden } from "@garden-io/sdk/testing"
 import { Log, PluginContext } from "@garden-io/sdk/types"
-import { getWorkspaces, setWorkspace } from "../common"
+import { getWorkspaces, setWorkspace } from "../helpers"
 import { expect } from "chai"
 import { defaultTerraformVersion, terraform } from "../cli"
