+++ plugins/terraform/test/terraform.ts
@@ -18,7 +18,7 @@ import { ConfigGraph, LogLevel } from "@garden-io/sdk/types"
 import { gardenPlugin } from ".."
 import { TerraformProvider } from "../provider"
 import { DeployTask } from "@garden-io/core/build/src/tasks/deploy"
-import { getWorkspaces, setWorkspace } from "../common"
+import { getWorkspaces, setWorkspace } from "../helpers"
 import { resolveAction } from "@garden-io/core/build/src/graph/actions"
 import { RunTask } from "@garden-io/core/build/src/tasks/run"
 import { defaultTerraformVersion } from "../cli"
